+++ src/detail/i_data.cpp
@@ -14,6 +14,7 @@
 #include <utility>
 #include <variant>
 
+#include <boost/numeric/conversion/cast.hpp>
 #include <boost/serialization/array.hpp>
 
 #if defined(HEYOKA_HAVE_REAL128)
@@ -28,7 +29,7 @@
 
 #endif
 
-#include <heyoka/detail/aligned_buffer.hpp>
+#include <heyoka/detail/aligned_vector.hpp>
 #include <heyoka/detail/i_data.hpp>
 #include <heyoka/detail/optional_s11n.hpp>
 #include <heyoka/detail/variant_s11n.hpp>
@@ -36,16 +37,14 @@
 #include <heyoka/s11n.hpp>
 #include <heyoka/taylor.hpp>
 
-// NOTE: implement a workaround for the serialisation of tuples whose first element
-// is a taylor outcome. We need this because Boost.Serialization treats all enums
-// as ints, which is not ok for taylor_outcome (whose underyling type will not
-// be an int on most platforms). Because it is not possible to override Boost's
-// enum implementation, we override the serialisation of tuples with outcomes
-// as first elements, which is all we need in the serialisation of the batch
+// NOTE: implement a workaround for the serialisation of tuples whose first element is a taylor outcome. We need this
+// because Boost.Serialization treats all enums as ints, which is not ok for taylor_outcome (whose underyling type will
+// not be an int on most platforms). Because it is not possible to override Boost's enum implementation, we override the
+// serialisation of tuples with outcomes as first elements, which is all we need in the serialisation of the batch
 // integrator. The implementation below will be preferred over the generic tuple
 // s11n because it is more specialised.
-// NOTE: this workaround is not necessary for the other enums in heyoka because
-// those all have ints as underlying type.
+//
+// NOTE: this workaround is not necessary for the other enums in heyoka because those all have ints as underlying type.
 namespace boost::serialization
 {
 
@@ -61,10 +60,8 @@ void save(Archive &ar, const std::tuple<heyoka::taylor_outcome, Args...> &tup, u
         }
     };
 
-    // NOTE: this is a right fold, which, in conjunction with the
-    // builtin comma operator, ensures that the serialisation of
-    // the tuple elements proceeds in the correct order and with
-    // the correct sequencing.
+    // NOTE: this is a right fold, which, in conjunction with the builtin comma operator, ensures that the serialisation
+    // of the tuple elements proceeds in the correct order and with the correct sequencing.
     std::apply([&tf](const auto &...x) { (tf(x), ...); }, tup);
 }
 
@@ -97,24 +94,35 @@ void serialize(Archive &ar, std::tuple<heyoka::taylor_outcome, Args...> &tup, un
 
 HEYOKA_BEGIN_NAMESPACE
 
-// Helper to initialise the compact-mode tape. Assumes an empty tape.
-template <typename T>
-void taylor_adaptive<T>::i_data::init_cm_tape()
+namespace detail
 {
-    assert(!m_tape);
 
-    const auto [sz, al] = m_tape_sa;
+namespace
+{
 
-    if (m_compact_mode) {
+// Common logic for initialising the compact-mode tape in scalar and batch integrators.
+template <typename IData>
+void init_cm_tape(IData &idata)
+{
+    assert(idata.m_cm_tape.empty());
+
+    const auto [sz, al] = idata.m_tape_sa;
+
+    if (idata.m_compact_mode) {
         assert(sz != 0u); // LCOV_EXCL_LINE
         assert(al != 0u); // LCOV_EXCL_LINE
 
-        m_tape = detail::make_aligned_buffer(sz, al);
+        idata.m_cm_tape = detail::aligned_vector<std::byte>(detail::aligned_allocator<std::byte>(al));
+        idata.m_cm_tape.resize(boost::numeric_cast<decltype(idata.m_cm_tape.size())>(sz));
     } else {
         assert(sz == 0u); // LCOV_EXCL_LINE
         assert(al == 0u); // LCOV_EXCL_LINE
     }
 }
+
+} // namespace
+
+} // namespace detail
 
 template <typename T>
 void taylor_adaptive<T>::i_data::save(boost::archive::binary_oarchive &ar, unsigned) const
@@ -160,21 +168,21 @@ void taylor_adaptive<T>::i_data::load(boost::archive::binary_iarchive &ar, unsig
     ar >> m_tm_data;
 
     // Recover the function pointers.
-    // NOTE: here we are recovering only the dense output function pointer because recovering
-    // the correct stepper requires information which is available only from the integrator
-    // class (hence, we do it from there).
+    //
+    // NOTE: here we are recovering only the dense output function pointer because recovering the correct stepper
+    // requires information which is available only from the integrator class (hence, we do it from there).
     m_d_out_f = std::visit([](auto &s) { return reinterpret_cast<d_out_f_t>(s.jit_lookup("d_out_f")); }, m_llvm_state);
 
     // Reconstruct the compact mode tape, if necessary.
-    m_tape.reset();
-    init_cm_tape();
+    m_cm_tape.clear();
+    detail::init_cm_tape(*this);
 }
 
-// NOTE: this ctor provides only partial initialisation of the data members.
-// The rest of the initialisation is performed from the integrator ctor.
-// NOTE: m_llvm_state is inited as a single llvm_state regardless of the use
-// of compact mode. It will be converted into a multi state if needed at a
-// later stage.
+// NOTE: this ctor provides only partial initialisation of the data members. The rest of the initialisation is performed
+// from the integrator ctor.
+//
+// NOTE: m_llvm_state is inited as a single llvm_state regardless of the use of compact mode. It will be converted into
+// a multi state if needed at a later stage.
 template <typename T>
 taylor_adaptive<T>::i_data::i_data(llvm_state s)
     : m_llvm_state(std::move(s)), m_tplt_state(std::get<0>(m_llvm_state).make_similar())
@@ -190,14 +198,15 @@ taylor_adaptive<T>::i_data::i_data(const i_data &other)
       m_tm_data(other.m_tm_data)
 {
     // Recover the function pointers.
-    // NOTE: here we are recovering only the dense output function pointer because recovering
-    // the correct stepper requires information which is available only from the integrator
-    // class (hence, we do it from there).
+    //
+    // NOTE: here we are recovering only the dense output function pointer because recovering the correct stepper
+    // requires information which is available only from the integrator class (hence, we do it from there).
+    //
     // NOLINTNEXTLINE(cppcoreguidelines-prefer-member-initializer)
     m_d_out_f = std::visit([](auto &s) { return reinterpret_cast<d_out_f_t>(s.jit_lookup("d_out_f")); }, m_llvm_state);
 
     // Init the compact mode tape, if necessary.
-    init_cm_tape();
+    detail::init_cm_tape(*this);
 }
 
 template <typename T>
@@ -226,25 +235,6 @@ HEYOKA_TAYLOR_ADAPTIVE_I_DATA_INST(mppp::real)
 #endif
 
 #undef HEYOKA_TAYLOR_ADAPTIVE_I_DATA_INST
-
-// Helper to initialise the compact-mode tape. Assumes an empty tape.
-template <typename T>
-void taylor_adaptive_batch<T>::i_data::init_cm_tape()
-{
-    assert(!m_tape);
-
-    const auto [sz, al] = m_tape_sa;
-
-    if (m_compact_mode) {
-        assert(sz != 0u); // LCOV_EXCL_LINE
-        assert(al != 0u); // LCOV_EXCL_LINE
-
-        m_tape = detail::make_aligned_buffer(sz, al);
-    } else {
-        assert(sz == 0u); // LCOV_EXCL_LINE
-        assert(al == 0u); // LCOV_EXCL_LINE
-    }
-}
 
 template <typename T>
 void taylor_adaptive_batch<T>::i_data::save(boost::archive::binary_oarchive &ar, unsigned) const
@@ -332,8 +322,8 @@ void taylor_adaptive_batch<T>::i_data::load(boost::archive::binary_iarchive &ar,
     m_d_out_f = std::visit([](auto &s) { return reinterpret_cast<d_out_f_t>(s.jit_lookup("d_out_f")); }, m_llvm_state);
 
     // Reconstruct the compact mode tape, if necessary.
-    m_tape.reset();
-    init_cm_tape();
+    m_cm_tape.clear();
+    detail::init_cm_tape(*this);
 }
 
 // NOTE: this ctor provides only partial initialisation of the data members.
@@ -369,7 +359,7 @@ taylor_adaptive_batch<T>::i_data::i_data(const i_data &other)
     m_d_out_f = std::visit([](auto &s) { return reinterpret_cast<d_out_f_t>(s.jit_lookup("d_out_f")); }, m_llvm_state);
 
     // Init the compact mode tape, if necessary.
-    init_cm_tape();
+    detail::init_cm_tape(*this);
 }
 
 template <typename T>